Rearrange the Odin Recipes Home page
Add margin and padding properties in order to rearrange the page and
center it like I would have wanted.

Add in a universal declaration of margin: 0; padding: 0; box-sizing:
border-box; in order to be in-line with current best practices.

Remember that margin: 0 auto; only works on block elements and when the
element in question has a width.