Add mismatched cards to two other links w/o flex
This is a test to create cards without using flexbox. The desired card
design at this point in time was achieved on the Chicken Noodle Soup
card.

Of course, the best and easiest way to get any of the results that I
have in my head will be to use flexbox, but it is still cool to know
that there is another way of creating a card.

This current layout sucks, though. However, this is all for the learning
experience.